LearningAutomata.jl API

Types

LearningAutomata.LRPLearner — Type

LRPLearner(n::Int,
           a::Vector{Float64},
           b::Vector{Float64}, 
           c::Vector{Float64},
           W::Vector{Float64},
           A::Matrix{Float64},
           R::Vector{Matrix{Float64}}, 
           P::Vector{Matrix{Float64}}) <: AbstractLinearLearner

A linear reward-penalty learner.

A general linear reward-penalty learner with n actions, learning rates a for rewards, learning rates b for punishments, action costs c, action probability vector W, vectors of reward and penalty operators R and P, and advantage matrix A.

References

Bush and Mosteller FIXME Kauhanen FIXME Narendra and Thathachar FIXME

Constructors

Functions

LearningAutomata.get_probs — Function

get_probs(x::AbstractLearner)

Return the current action probability vector.

LearningAutomata.reward! — Function

reward!(x::AbstractLinearLearner, i::Int)

Reward the ith action of an AbstractLinearLearner.

LearningAutomata.punish! — Function

punish!(x::AbstractLinearLearner, i::Int)

Punish the ith action of an AbstractLinearLearner.

LearningAutomata.interact! — Function

interact!(x::AbstractLearner, y::AbstractLearner; symmetric = true)

Make two learners x and y interact.

Interaction is symmetric, i.e. both learners update their state, if symmetric = true. Else it is asymmetric, and only y updates its state.

LearningAutomata.simulate! — Function

simulate!(x::AbstractLearner, iter::Int, c::Vector{Float64})

Simulate a learner for iter iterations in a stationary random environment constituted by penalty probability vector c.

Returns the trajectory as a $m \times n$ matrix, where $m$ is the trajectory length (number of simulation iterations) and $n$ is the learner's dimensionality (number of actions). This can be redirected e.g. to plot in order to visualize the learning trajectory.